2.2.StreamLit_chatbot/requirements.txt
streamlit==1.34.0
sqlmodel==0.0.18
streamlit-authenticator==0.3.2
streamlit-option-menu==0.3.12
streamlit-cookie==0.1.0
groq==0.5.0
psycopg2-binary
python-dotenv
